services/context_adapter/requirements.txt
fastapi
pydantic
uvicorn[standard]
asyncio-mqtt
jsonschema
PyYAML
